+++ src/app/control-and-view/employee/trade-request-view/trade-request-view.component.ts
@@ -2,6 +2,9 @@ import { Component, OnInit } from '@angular/core';
 import { PeopleServiceService } from "../../../service/people-service.service";
 
 import { DataServiceTokenStorageService } from '../../../service/DataServiceTokenStorage.service';
+import { MatDialog } from '@angular/material/dialog';
+import { AlertdialogComponent } from '../../dialog/alertdialog/alertdialog.component';
+import { ConfirmationdialogComponent, ConfirmDialogModel } from '../../dialog/confirmationdialog/confirmationdialog.component';
 @Component({
   selector: 'app-trade-request-view',
   templateUrl: './trade-request-view.component.html',
@@ -39,23 +42,68 @@ export class TradeRequestViewComponent implements OnInit {
     return window.atob(output);
   }
 
-  constructor(private PeopleServiceService: PeopleServiceService, private dst: DataServiceTokenStorageService) { }
+  constructor(private PeopleServiceService: PeopleServiceService, private dst: DataServiceTokenStorageService, private dialog: MatDialog) { }
 
   deletePass(key) {
     this.deleteRequestKey = key;
+
+
+    const message = `Are you sure !!  Do you want to delete`;
+    const dialogData = new ConfirmDialogModel("DELETE", message);
+    const dialogRef = this.dialog.open(ConfirmationdialogComponent, {
+      maxWidth: "400px",
+      data: dialogData
+    });
+
+    dialogRef.afterClosed().subscribe(dialogResult => {
+      if (dialogResult) {
+
+        this.checkFlag = true;
+        this.PeopleServiceService.deleteTradeRequest(this.deleteRequestKey, this.toServeremployeekey)
+          .subscribe((data) => {
+            this.checkFlag = false;
+            // alert('Trade Request Deleted Successfully');
+            const dialogRef = this.dialog.open(AlertdialogComponent, {
+              data: {
+                message: 'Trade Request Deleted Successfully',
+                buttonText: {
+                  cancel: 'Done'
+                }
+              },
+            });
+            dialogRef.afterClosed().subscribe(dialogResult => {
+              this.PeopleServiceService.getTradeRequestdetails(this.OrganizationID, this.toServeremployeekey).subscribe((data) => {
+                this.requestdetails = data;
+              });
+            });
+          });
+      } else {
+        this.checkFlag = false;
+      }
+    });
   }
 
-  deleteRequest() {
-    this.checkFlag = true;
-    this.PeopleServiceService.deleteTradeRequest(this.deleteRequestKey, this.toServeremployeekey)
-      .subscribe((data) => {
-        this.checkFlag = false;
-        alert('Trade Request Deleted Successfully');
-        this.PeopleServiceService.getTradeRequestdetails(this.OrganizationID, this.toServeremployeekey).subscribe((data) => {
-          this.requestdetails = data;
-        });
-      });
-  }
+  // deleteRequest() {
+  //   this.checkFlag = true;
+  //   this.PeopleServiceService.deleteTradeRequest(this.deleteRequestKey, this.toServeremployeekey)
+  //     .subscribe((data) => {
+  //       this.checkFlag = false;
+  //       // alert('Trade Request Deleted Successfully');
+  //       const dialogRef = this.dialog.open(AlertdialogComponent, {
+  //         data: {
+  //           message: 'Trade Request Deleted Successfully',
+  //           buttonText: {
+  //             cancel: 'Done'
+  //           }
+  //         },
+  //       });
+  //       dialogRef.afterClosed().subscribe(dialogResult => {
+  //         this.PeopleServiceService.getTradeRequestdetails(this.OrganizationID, this.toServeremployeekey).subscribe((data) => {
+  //           this.requestdetails = data;
+  //         });
+  //       });
+  //     });
+  // }
   ngOnInit() {
 
     // var token = sessionStorage.getItem('token');
@@ -66,7 +114,7 @@ export class TradeRequestViewComponent implements OnInit {
     this.name = this.dst.getName();
     this.toServeremployeekey = this.dst.getEmployeekey();
     this.OrganizationID = this.dst.getOrganizationID();
-    
+
     this.checkFlag = false;
 
     this.checkEmp = parseInt(this.toServeremployeekey);
